Update package name to webpack-config
