+++ docs/manifest.json
@@ -834,6 +834,12 @@
 		"parent": "packages"
 	},
 	{
+		"title": "@wordpress/webpack-config",
+		"slug": "packages-webpack-config",
+		"markdown_source": "https://raw.githubusercontent.com/WordPress/gutenberg/master/packages/webpack-config/README.md",
+		"parent": "packages"
+	},
+	{
 		"title": "@wordpress/wordcount",
 		"slug": "packages-wordcount",
 		"markdown_source": "https://raw.githubusercontent.com/WordPress/gutenberg/master/packages/wordcount/README.md",
