--- packages/build-config/README.md
@@ -1,2 +1,0 @@
-# Build Config
-
