--- packages/build-config/index.js
@@ -1,6 +1,0 @@
-const webpackConfig = require( './config/webpack.config.js' );
-
-module.exports = {
-	webpackConfig,
-};
-
