--- packages/build-config/package.json
@@ -1,88 +1,0 @@
-{
-	"name": "@wordpress/build-config",
-	"version": "1.0.0",
-	"description": "WordPress JavaScript build configuration.",
-	"author": "The WordPress Contributors",
-	"license": "GPL-2.0-or-later",
-	"keywords": [
-		"wordpress",
-		"build",
-		"config"
-	],
-	"homepage": "https://github.com/WordPress/gutenberg/tree/master/packages/build-config/README.md",
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/WordPress/gutenberg.git"
-	},
-	"bugs": {
-		"url": "https://github.com/WordPress/gutenberg/issues"
-	},
-	"engines": {
-		"node": ">=8",
-		"npm": ">=6.0.0"
-	},
-	"files": [
-		"build",
-		"build-module"
-	],
-	"main": "build/index.js",
-	"module": "build-module/index.js",
-	"dependencies": {
-		"@babel/core": "7.0.0",
-		"@babel/runtime-corejs2": "7.0.0",
-		"@wordpress/babel-plugin-import-jsx-pragma": "file:../babel-plugin-import-jsx-pragma",
-		"@wordpress/babel-plugin-makepot": "file:../babel-plugin-makepot",
-		"@wordpress/babel-preset-default": "file:../babel-preset-default",
-		"@wordpress/browserslist-config": "file:../browserslist-config",
-		"@wordpress/custom-templated-path-webpack-plugin": "file:../custom-templated-path-webpack-plugin",
-		"@wordpress/eslint-config": "file:../eslint-config",
-		"@wordpress/jest-console": "file:../jest-console",
-		"@wordpress/jest-preset-default": "file:../jest-preset-default",
-		"@wordpress/library-export-default-webpack-plugin": "file:../library-export-default-webpack-plugin",
-		"@wordpress/npm-package-json-lint-config": "file:../npm-package-json-lint-config",
-		"@wordpress/postcss-themes": "file:../postcss-themes",
-		"autoprefixer": "8.2.0",
-		"babel-loader": "8.0.0",
-		"chalk": "2.4.1",
-		"check-node-version": "3.1.1",
-		"concurrently": "3.5.0",
-		"core-js": "2.5.7",
-		"cross-env": "3.2.4",
-		"deasync": "0.1.13",
-		"deep-freeze": "0.0.1",
-		"doctrine": "2.1.0",
-		"eslint-plugin-jest": "21.5.0",
-		"espree": "3.5.4",
-		"glob": "7.1.2",
-		"husky": "0.14.3",
-		"jest-puppeteer": "3.2.1",
-		"lerna": "3.4.3",
-		"lint-staged": "7.2.0",
-		"lodash": "4.17.10",
-		"mkdirp": "0.5.1",
-		"node-sass": "4.9.2",
-		"path-type": "3.0.0",
-		"pegjs": "0.10.0",
-		"phpegjs": "1.0.0-beta7",
-		"postcss-color-function": "4.0.1",
-		"puppeteer": "1.6.1",
-		"react-test-renderer": "16.6.3",
-		"rimraf": "2.6.2",
-		"rtlcss": "2.4.0",
-		"sass-loader": "6.0.7",
-		"source-map-loader": "0.2.3",
-		"sprintf-js": "1.1.1",
-		"stylelint": "9.5.0",
-		"stylelint-config-wordpress": "13.1.0",
-		"symlink-or-copy": "1.2.0",
-		"uuid": "3.3.2",
-		"webpack": "4.8.3",
-		"webpack-bundle-analyzer": "3.0.2",
-		"webpack-cli": "2.1.3",
-		"webpack-livereload-plugin": "2.1.1",
-		"webpack-rtl-plugin": "github:yoavf/webpack-rtl-plugin#develop"
-	},
-	"publishConfig": {
-		"access": "public"
-	}
-}
