+++ packages/webpack-config/README.md
@@ -1,0 +1,13 @@
+# Webpack Config
+
+[Webpack](https://webpack.js.org/) config for WordPress development.
+
+## Installation
+
+Install the module
+
+```bash
+npm install @wordpress/webpack-config --save-dev
+```
+
+<br/><br/><p align="center"><img src="https://s.w.org/style/images/codeispoetry.png?1" alt="Code is Poetry." /></p>
