+++ webpack.config.js
@@ -1,14 +1,19 @@
 /**
  * External dependencies
  */
+const WebpackRTLPlugin = require( 'webpack-rtl-plugin' );
 const CopyWebpackPlugin = require( 'copy-webpack-plugin' );
 const postcss = require( 'postcss' );
+
+const { get } = require( 'lodash' );
+const { basename } = require( 'path' );
 
 /**
  * WordPress dependencies
  */
-
-const webpackConfig = require( '@wordpress/build-config' ).webpackConfig;
+const CustomTemplatedPathPlugin = require( '@wordpress/custom-templated-path-webpack-plugin' );
+const LibraryExportDefaultPlugin = require( '@wordpress/library-export-default-webpack-plugin' );
+const webpackConfig = require( '@wordpress/webpack-config' );
 
 /**
  * Internal dependencies
@@ -45,6 +50,40 @@ webpackConfig.entry = gutenbergPackages.reduce( ( memo, packageName ) => {
 }, {} );
 
 webpackConfig.plugins.push(
+	// Create RTL files with a -rtl suffix
+	new WebpackRTLPlugin( {
+		suffix: '-rtl',
+		minify: process.env.NODE_ENV === 'production' ? { safe: true } : false,
+	} ),
+	new CustomTemplatedPathPlugin( {
+		basename( path, data ) {
+			let rawRequest;
+
+			const entryModule = get( data, [ 'chunk', 'entryModule' ], {} );
+			switch ( entryModule.type ) {
+				case 'javascript/auto':
+					rawRequest = entryModule.rawRequest;
+					break;
+
+				case 'javascript/esm':
+					rawRequest = entryModule.rootModule.rawRequest;
+					break;
+			}
+
+			if ( rawRequest ) {
+				return basename( rawRequest );
+			}
+
+			return path;
+		},
+	} ),
+	new LibraryExportDefaultPlugin( [
+		'api-fetch',
+		'deprecated',
+		'dom-ready',
+		'redux-routine',
+		'token-list',
+	].map( camelCaseDash ) ),
 	new CopyWebpackPlugin(
 		gutenbergPackages.map( ( packageName ) => ( {
 			from: `./packages/${ packageName }/build-style/*.css`,
